social_bot/requirements.txt
instabot
textblob
python-dotenv